Rename WifiHotspotHelper to WifiHotspotConnector
