+++ skunkworks_crow/src/main/java/org/odk/share/activities/SendActivity.java
@@ -19,10 +19,10 @@ import android.widget.TextView;
 import android.widget.Toast;
 
 import org.odk.share.R;
-import org.odk.share.controller.WifiHotspotHelper;
 import org.odk.share.events.HotspotEvent;
 import org.odk.share.events.UploadEvent;
 import org.odk.share.network.WifiConnector;
+import org.odk.share.network.WifiHospotConnector;
 import org.odk.share.rx.RxEventBus;
 import org.odk.share.rx.schedulers.BaseSchedulerProvider;
 import org.odk.share.services.HotspotService;
@@ -59,7 +59,7 @@ public class SendActivity extends InjectableActivity {
     @Inject
     BaseSchedulerProvider schedulerProvider;
     @Inject
-    WifiHotspotHelper wifiHotspot;
+    WifiHospotConnector wifiHotspot;
     @Inject
     SenderService senderService;
