+++ skunkworks_crow/src/main/java/org/odk/share/injection/config/AppModule.java
@@ -6,10 +6,10 @@ import android.content.Context;
 
 import org.odk.collect.android.dao.FormsDao;
 import org.odk.collect.android.dao.InstancesDao;
-import org.odk.share.controller.WifiHotspotHelper;
 import org.odk.share.dao.InstanceMapDao;
 import org.odk.share.dao.TransferDao;
 import org.odk.share.injection.config.scopes.PerApplication;
+import org.odk.share.network.WifiHospotConnector;
 import org.odk.share.rx.RxEventBus;
 import org.odk.share.rx.schedulers.BaseSchedulerProvider;
 import org.odk.share.rx.schedulers.SchedulerProvider;
@@ -58,8 +58,8 @@ class AppModule {
 
     @Provides
     @PerApplication
-    WifiHotspotHelper provideHotspotHelper(Context context) {
-        return new WifiHotspotHelper(context);
+    WifiHospotConnector provideHotspotHelper(Context context) {
+        return new WifiHospotConnector(context);
     }
 
     @Provides
